# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Japan SSW Platform API",
    "version": "1.0.0",
    "description": "REST API for Japan Specified Skilled Worker (SSW) job platform - connects Filipino jobseekers with Japanese employers",
    "contact": {
      "name": "API Support",
      "email": "support@japanssw.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.japanssw.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login"
      }
    },
...
  },
...
  "paths": {
    "/api/v1/jobs/{jobId}/apply": {
      "post": {
        "summary": "Apply to a job",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverLetter": {
                    "type": "string",
                    "maxLength": 2000,
                    "example": "I am very interested in this position..."
                  },
                  "resumePath": {
                    "type": "string",
                    "example": "/uploads/resumes/resume.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted"
          },
          "400": {
            "description": "Already applied or job closed"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/v1/applications/me": {
      "get": {
        "summary": "Get current user's applications",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "submitted",
                "reviewing",
                "interview",
                "offer",
                "accepted",
                "rejected",
                "withdrawn"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications retrieved successfully"
          }
        }
      }
    },
    "/api/v1/applications/{id}": {
      "get": {
        "summary": "Get single application details",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application retrieved"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Application not found"
          }
        }
      }
    },
    "/api/v1/applications/{id}/withdraw": {
      "put": {
        "summary": "Withdraw application",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application withdrawn successfully"
          },
          "400": {
            "description": "Cannot withdraw application in current status"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/applications/{id}/status": {
      "put": {
        "summary": "Update application status (Employer/Admin)",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "submitted",
                      "reviewing",
                      "interview",
                      "offer",
                      "accepted",
                      "rejected"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  },
                  "interview": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "location": {
                        "type": "string"
                      },
                      "notes": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/applications/{id}/notes": {
      "put": {
        "summary": "Add employer notes to application",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "notes"
                ],
                "properties": {
                  "notes": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notes added successfully"
          }
        }
      }
    },
...
    "/api/v1/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "japaneseLevel": {
                    "type": "string",
                    "enum": [
                      "N5",
                      "N4",
                      "N3",
                      "N2",
                      "N1",
                      "Native"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created"
          }
        }
      },
      "put": {
        "summary": "Update user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated"
          }
        }
      },
      "delete": {
        "summary": "Delete user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile deleted"
          }
        }
      }
    },
...
  }
}